Encapsulate Penn-specific code and data.
To do so, this commit introduces the local/ directory, which contains a
subdirectory for each institution, and config.py to choose between them.
fetch_all_events() now calls every public function in the config.scrapers
module and concatenates the results. (The fetch_* functions have been
moved to local/upenn/scrapers.py, and config.scrapers points at that).
Likewise, create_event_features now uses the locations listed in
config.location (which now points at local/upenn/locations.py)

A few other configution options (paths, grobid URLs) were moved to config.py.
